remove test code from non-test dir
